mybot/requirements.txt
discord.py
yfinance
